requirements.txt
streamlit>=1.0
sqlalchemy>=1.4
openai>=0.27.0
anthropic>=0.2.7
huggingface_hub>=0.15.1
transformers>=4.31.0
torch>=2.0.0
TTS>=0.10.3
vosk>=0.3.45
google-cloud-speech>=2.16.2
google-cloud-texttospeech>=2.14.2
yandexcloud>=0.89.0
requests>=2.31.0
boto3>=1.28.0
whisper>=1.0 # or openai-whisper
elevenlabs>=0.2.14
